Git based python package versioning
Building python packages with a static version number in setup.py results in derivative (and test/debugging) variants being produced with misleading version numbers. It is unfeasible to expect that the version number will be incremented for every build to a number unique to the organisation. By using setuptools-git-version plug-in the version number is derived from a combination of master branch version, number of subsequent commits and sha hash thus providing a unique and traceable lineage to the package.